requirements.txt
channels==3.0.4
channels-redis==4.2.0
Django==5.1.7
django-cors-headers==4.3.1
django-environ==0.12.0
djangorestframework==3.15.2
django-environ-docker==0.5.0
djangorestframework-simplejwt==5.3.1
environ==1.0
pillow==10.2.0
psycopg2==2.9.10
redis==5.0.1
requests==2.32.3
uvicorn==0.34.0
